Renamed historicalChart component dir as historicalData
