+++ src/pages/details/ocpCloudDetails/detailsTableItem.tsx
@@ -10,7 +10,7 @@ import {
 import { ReportPathsType } from 'api/reports/report';
 import { BulletChart } from 'pages/details/components/bulletChart/bulletChart';
 import { Cluster } from 'pages/details/components/cluster/cluster';
-import { HistoricalModal } from 'pages/details/components/historicalChart/historicalModal';
+import { HistoricalModal } from 'pages/details/components/historicalData/historicalModal';
 import { Tag } from 'pages/details/components/tag/tag';
 import React from 'react';
 import { InjectedTranslateProps, translate } from 'react-i18next';
